finish memory allocator

+++ src/string.c
@@ -8,7 +8,7 @@ size_t strlen(const char *s) {
     return i;
 }
 
-char *itoa(int n, char *str, int radix) {
+char *itoa(u64 n, char *str, int radix) {
     static char digits[] = "0123456789abcdefghijklmnopqrstuvwxyz";
     char *p = str;
     char *head = str;
